% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prostate_redcap.R
\name{check_prostate_redcap}
\alias{check_prostate_redcap}
\title{Quality Control Check for Prostate REDCap database}
\usage{
check_prostate_redcap(
  data,
  qc_crit_pts = qc_criteria_pts(),
  qc_crit_smp = qc_criteria_smp(),
  qc_level_pts = NULL,
  qc_level_smp = NULL,
  recommended_only = FALSE
)
}
\arguments{
\item{data}{List with elements \code{pts} and \code{smp},
returned by \code{\link{load_prostate_redcap}}.}

\item{qc_crit_pts}{Criteria for checking the \code{pts} tibble.
Defaults to the return of the \code{\link{qc_criteria_pts}} function.
Custom criteria can be supplied instead.}

\item{qc_crit_smp}{Criteria for checking the \code{smp} tibble.
Defaults to the return of the \code{\link{qc_criteria_smp}} function.
Custom criteria can be supplied instead.}

\item{qc_level_pts}{Level of QC that the return \code{pts} tibble
will be restricted to. By default, all QC steps in \code{qc_crit_pts}
will be applied. A integer index (row) number of
\code{qc_crit_pts} can be provided instead to perform
less strict exclusions. \code{qc_level_pts = 1} will perform
no exclusions.}

\item{qc_level_smp}{As \code{qc_level_pts}, for \code{smp}.}

\item{recommended_only}{Return qc'd \code{pts} and \code{smp}
restricted to variables that are recommended for use in analyses?
Defaults to \code{FALSE} but is recommended for use.}
}
\value{
List:
\itemize{
\item \code{pts}: Patient-level data after QC.
\item \code{smp}: Sample-level data after QC.
\item \code{qc_pts}: Tibble of sequential exclusions for \code{pts}.
\item \code{qc_smp}: Tibble of sequential exclusions for \code{smp}.
The \code{qc_pts} and \code{qc_smp} tibbles can be used to extract
information on which records failed which QC step.
}
}
\description{
Run sequential quality control checks and
thereby restrict the \code{pts} and \code{smp} datasets
to cases that pass certain criteria.
}
\examples{
\dontrun{
# Process output of load_prostate_redcap():
pts_smp_qc <- check_prostate_redcap(pts_smp, recommended_only = TRUE)
}
}
\seealso{
Overview of analysis-ready data elements:
\url{https://stopsack.github.io/prostateredcap/reference/dataelements.html}
}
