% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify_prostate_redcap.R
\name{deidentify_prostate_redcap}
\alias{deidentify_prostate_redcap}
\title{Deidentify Prostate REDCap Data}
\usage{
deidentify_prostate_redcap(data, trt_freetext = TRUE, extra = NULL)
}
\arguments{
\item{data}{List with elements \code{pts}, \code{smp}, and \code{trt},
as generated by \code{\link{load_prostate_redcap}}.}

\item{trt_freetext}{Optional. Remove free text treatment names? Defaults to
\code{TRUE}.}

\item{extra}{Optional. Tibble with additional data set, or list of additional
data sets, in which a \code{ptid} variable of type \code{character} should
be replaced with the same sequential ID as in \code{pts}. Other
deidentification steps of such extra datasets must be done separately if
necessary. Defaults to \code{NULL} (none).}
}
\value{
List:
\itemize{
\item \code{pts}: Deidentified patient-level data.
\item \code{smp}: Deidentified sample-level data.
\item \code{trt}: Deidentified treatment data.
\item \code{ext}: List of additional data sets (if provided, otherwise
empty, \code{NULL}).
}
}
\description{
Deidentifies the dataset for analysis and sharing:
\itemize{
\item Drops all dates, keeping time intervals only
\item Rounds age (in years, 1 decimal) and PSA (1 decimal)
\item Replaces patient ID by a sequential index number that still allows for
merging \code{pts}, \code{smp}, and \code{trt} datasets.
\item Removes free text with potentially identifying information (treatments).
}

By default, this function is already being called automatically within
\code{\link{load_prostate_redcap}}.
}
