% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prostate_redcap.R
\name{guessdate}
\alias{guessdate}
\title{Guess Dates}
\usage{
guessdate(messydate)
}
\arguments{
\item{messydate}{Date in the format MM/DD/YYYY,
MM/YYYY, or YYYY only.}
}
\value{
Date vector in ISO date notation.
}
\description{
Uses \code{\link[lubridate]{parse_date_time}}
to transform string vectors with different date formats into
an R date vector, printed in ISO format.
If only the year is provided, then
the date will be set as June 30 of that year.
}
\examples{
guessdate(c("03/28/2019", "03/2019", "03/19", "2019"))
}
