% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prostate_redcap.R
\name{load_prostate_redcap}
\alias{load_prostate_redcap}
\title{Load MSK-IMPACT Prostate REDCap Labelled CSV File}
\usage{
load_prostate_redcap(labelled_csv, deidentify = TRUE)
}
\arguments{
\item{labelled_csv}{CSV file with labels, exported from REDCap.
Must be the labelled version and must contain dates
in order to derive time intervals.}

\item{deidentify}{De-identify the returned data set?
Drops all dates (keeping time intervals), rounds age and PSA,
and replaces patient ID by an index number. Defaults to
\code{TRUE} and should rarely, if ever, be turned off.}
}
\value{
List of two labelled tibbles (data frames):
\itemize{
\item \code{pts}: Patient-level data
\item \code{smp}: Sample-level data
}

Access variables labels in RStudio via \code{\link[utils]{View}}
or using \code{\link{attr}(., "label")}.
}
\description{
Loads, merges, reformats, corrects, and labels
the REDCap file used for the MSK-IMPACT Prostate clinical database.
It is recommended that the returned list is next processed by
\code{\link{check_prostate_redcap}}.
}
\details{
The following edits and assumptions are made:
\enumerate{
\item Potentially incomplete date variables are converted to
date vectors, using \code{\link{guessdate}}.
\item Various missingness indicators in strings and factors,
\code{c("Unknown / Not Reported", "N/A", "NA", "Unknown", "X", "x")},
are converted to \code{NA}.
\item "Undetectable" PSA is set to 0, PSA \code{">x"} is set to \code{x + 1}.
\item Clinical T and N stage variables are set to missing if M1.
\item Event dates for metastases (\code{met_date}) and castration resistance
(\code{crpc_date}) are set:
\itemize{
\item To the last clinic visit (\code{lastvisit})
if the event has not occurred.
\item If stage is M1 and the recorded metastasis date is no more than
1 month discrepant, \code{met_date} is set to the diagnosis
date (\code{dxdate}).
\item If the sample is a variant histology (e.g., neuroendocrine),
the castration resistance date (\code{crpc_date}) is the date of
diagnosis and the event indicator for survival analyses
(\code{event_crpc}) is \code{NA}.
}
\item Disease extent, distinguishing CRPC from castration-sensitive disease,
at sampling is based on the sample date and the date of the sample.
If the samples was obtained before the CRPC date, or CRPC did not occur,
the the sample is from castration-sensitive disease by definition.
}
}
\examples{
\dontrun{
# Load data:
pts_smp <- load_prostate_redcap(
  labelled_csv = "GUPIMPACTDatabaseFre_DATA_LABELS_2020-01-01_0001.csv")

# Access patient-level data:
pts_smp$pts

# Access sample-level data:
pts_smp$smp

# Pass 'pts_smp' to check_prostate_redcap() next

}
}
