% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostateredcap.R
\docType{package}
\name{prostateredcap}
\alias{prostateredcap}
\alias{prostateredcap-package}
\title{Load and check the MSK Prostate REDCap Prostate clinical database}
\description{
This package loads, merges, reformats, corrects, labels, and quality controls
the MSK Prostate REDCap Prostate clinical database.
Run \code{\link{load_prostate_redcap}}, followed by \code{\link{check_prostate_redcap}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stopsack.github.io/prostateredcap}
  \item Report bugs at \url{https://github.com/stopsack/prostateredcap/issues}
}

}
\author{
\strong{Maintainer}: Konrad Stopsack \email{stopsack@post.harvard.edu} (\href{https://orcid.org/0000-0002-0722-1311}{ORCID})

}
\keyword{internal}
