% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prostate_redcap.R
\name{qc_criteria_pts}
\alias{qc_criteria_pts}
\title{Quality Control Criteria for Patient-Level Data}
\usage{
qc_criteria_pts()
}
\value{
Tibble:
\itemize{
\item \code{label}: Description of the exclusion criterion.
\item \code{filter_criterion}: Expression that will be used in
\code{\link[dplyr]{filter}}
for restriction to observations that pass the criterion.
}
}
\description{
For use in \code{\link{check_prostate_redcap}}.
}
