% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prostate_redcap.R
\name{load_prostate_redcap}
\alias{load_prostate_redcap}
\title{Load MSK-IMPACT Prostate REDCap Labeled CSV File}
\usage{
load_prostate_redcap(
  labeled_csv,
  deidentify = TRUE,
  keep_also = list(baseline = NULL, sample = NULL, freeze = NULL)
)
}
\arguments{
\item{labeled_csv}{CSV file with labels, exported from REDCap.
Must be the labeled version and must contain dates
in order to derive time intervals.}

\item{deidentify}{De-identify the returned data set using
\code{\link{deidentify_prostate_redcap}}? Defaults to
\code{TRUE}. Should only be disabled if additional data need
to be merged by identifiers, followed by calling
\code{\link{deidentify_prostate_redcap}} separately.}

\item{keep_also}{Optional. Additional patient-level variables to keep
without editing. As applicable, they would need to be
deidentified manually.
Provide as list with vectors of variable names for baseline and freeze
forms: \code{list(baseline = c("var1", "var2"), freeze = "varX")}.}
}
\value{
List of two labeled tibbles (data frames):
\itemize{
\item \code{pts}: Patient-level data
\item \code{smp}: Sample-level data
}

Access variables labels in RStudio via \code{\link[utils]{View}}
or using \code{\link{attr}(., "label")}.

The warning message, \code{Duplicated column names deduplicated}, is
expected due to the design of the REDCap dataset.
}
\description{
Loads, merges, reformats, corrects, and labels
the REDCap file used for the MSK-IMPACT Prostate clinical database.
It is recommended that the returned list is next processed by
\code{\link{check_prostate_redcap}}.
}
\details{
The following edits and assumptions are made:
\enumerate{
\item Potentially incomplete date variables are converted to
date format, using \code{\link{guessdate}}.
\item Various missingness indicators in strings and factors,
\code{c("Unknown / Not Reported", "N/A", "NA", "Unknown", "X", "x")},
are converted to \code{NA}.
\item "Undetectable" PSA is set to 0, PSA \code{">x"} is set to \code{x + 1},
PSA \code{"a-b"} (e.g., \code{4.5-4.7}) is set to the mean of the two
values.
\item Clinical T and N stage variables are set to missing if M1.
\item Event dates and follow-up time for metastases (\code{met_date}),
castration resistance (\code{crpc_date}), and death are set:
\itemize{
\item Event date is the last clinic visit (\code{lastvisit})
if a CRPC/metastases event has not occurred.
\item Event date is the last follow up/contact (\code{lastfu})
if last known survival status is alive.
\item If stage is M1 and the recorded metastasis date is no more than
1 month discrepant, \code{met_date} is set to the diagnosis
date (\code{dxdate}).
\item If the sample is a variant histology (e.g., neuroendocrine),
the castration resistance date (\code{crpc_date}) is the date of
diagnosis and the event indicator for survival analyses
(\code{event_crpc}) is \code{NA}.
\item Time intervals for these three survival outcomes are calculated
from the time of sequencing. For late-entry survival models,
time intervals from diagnosis to sequencing and from sample/biopsy to
sequencing are also provided.
}
\item Disease extent, distinguishing CRPC from castration-sensitive disease,
at sampling is based on the sample date and the date of castration
resistance. If the samples was obtained before the CRPC date, or
CRPC did not occur, the sample is from castration-sensitive disease
by definition.
}
}
\examples{
\dontrun{
# Load data:
pts_smp <- load_prostate_redcap(
  labeled_csv = "GUPIMPACTDatabaseFre_DATA_LABELS_2020-01-01_0001.csv")

# Access patient-level data:
pts_smp$pts

# Access sample-level data:
pts_smp$smp

# Pass 'pts_smp' to check_prostate_redcap() next
}
}
\seealso{
Overview of analysis-ready data elements:
\url{https://stopsack.github.io/prostateredcap/articles/dataelements.html}
}
