% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prostate_redcap.R
\name{qc_criteria_smp}
\alias{qc_criteria_smp}
\title{Quality Control Criteria for Sample-Level Data}
\usage{
qc_criteria_smp()
}
\value{
Tibble:
\itemize{
\item \code{label}: Description of the exclusion criterion.
\item \code{filter_criterion}: Expression that will be used in
\code{\link[dplyr]{filter}}
for restriction to observations that pass the criterion.
\item \code{index}: Level of QC that can be provided to
\code{\link{check_prostate_redcap}} in the \code{qc_level_pts} and
\code{qc_level_smp} arguments.
}
}
\description{
For use in \code{\link{check_prostate_redcap}}.
}
\examples{
# Show all criteria:
qc_criteria_smp()

# Show code for criterion #4:
qc_criteria_smp()$filter_criterion[[4]]
}
